+++ src/main/java/ru/danon/spring/ToDo/repositories/jpa/ScheduledNotificationRepository.java
@@ -8,18 +8,18 @@ import ru.danon.spring.ToDo.models.postgre.ScheduledNotification;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface ScheduledNotificationRepository extends JpaRepository<ScheduledNotification, Integer> {
+public interface ScheduledNotificationRepository extends JpaRepository<ScheduledNotification, Long> {
     // Для обработки: находим уведомления, готовые к отправке
     List<ScheduledNotification> findByStatusAndScheduledTimeBetween(
             String status, LocalDateTime start, LocalDateTime end);
 
     // Для отмены: находим все pending уведомления для задачи
     List<ScheduledNotification> findByTaskIdAndUserIdAndStatus(
-            Integer taskId, Integer userId, String status);
+            Long taskId, Long userId, String status);
 
     // Проверяем, не было ли уже запланировано такое уведомление
     boolean existsByTaskIdAndUserIdAndEventTypeAndStatus(
-            Integer taskId, Integer userId, String eventType, String status);
+            Long taskId, Long userId, String eventType, String status);
 
     // Для повторной обработки failed уведомлений
     List<ScheduledNotification> findByStatusAndAttemptCountLessThan(
@@ -27,8 +27,8 @@ public interface ScheduledNotificationRepository extends JpaRepository<Scheduled
 
     // Находим уведомления по задаче и пользователю
     @Query("SELECT sn FROM ScheduledNotification sn WHERE sn.taskId = :taskId AND sn.userId = :userId")
-    List<ScheduledNotification> findByTaskAndUser(@Param("taskId") Integer taskId,
-                                                  @Param("userId") Integer userId);
+    List<ScheduledNotification> findByTaskAndUser(@Param("taskId") Long taskId,
+                                                  @Param("userId") Long userId);
 
-    List<ScheduledNotification> findByTaskId(Integer taskId);
+    List<ScheduledNotification> findByTaskId(Long taskId);
 }
