+++ src/main/java/ru/danon/spring/ToDo/services/impl/NotificationSchedulingServiceImpl.java
@@ -23,7 +23,9 @@ public class NotificationSchedulingServiceImpl implements NotificationScheduling
     // Конфигурируемые интервалы уведомлений (часы до дедлайна)
     private final List<Integer> NOTIFICATION_INTERVALS = Arrays.asList(48, 24, 12);
 
-    @Transactional @Override public void scheduleTaskNotifications(TaskAssignment assignment) {
+    @Transactional
+    @Override
+    public void scheduleTaskNotifications(TaskAssignment assignment) {
         LocalDateTime deadline = assignment.getTask().getDeadline();
         if (deadline == null) {
             log.warn("Задача id={} не имеет дедлайна, пропускаем планирование уведомлений", assignment.getTask().getId());
@@ -40,7 +42,9 @@ public class NotificationSchedulingServiceImpl implements NotificationScheduling
         }
     }
 
-    @Transactional @Override public void rescheduleTaskNotifications(TaskAssignment assignment) {
+    @Transactional
+    @Override
+    public void rescheduleTaskNotifications(TaskAssignment assignment) {
         log.info("Перепланирование уведомлений для задачи id={}, пользователь id={}",
                 assignment.getTask().getId(), assignment.getUserId());
 
@@ -51,7 +55,9 @@ public class NotificationSchedulingServiceImpl implements NotificationScheduling
         scheduleTaskNotifications(assignment);
     }
 
-    @Transactional @Override public void cancelTaskNotifications(Integer taskId, Integer userId) {
+    @Transactional
+    @Override
+    public void cancelTaskNotifications(Long taskId, Long userId) {
         List<ScheduledNotification> pendingNotifications =
                 scheduledNotificationRepository.findByTaskIdAndUserIdAndStatus(taskId, userId, "PENDING");
 
@@ -65,7 +71,9 @@ public class NotificationSchedulingServiceImpl implements NotificationScheduling
         }
     }
 
-    @Transactional @Override public void cancelAllTaskNotifications(Integer taskId) {
+    @Transactional
+    @Override
+    public void cancelAllTaskNotifications(Long taskId) {
         List<ScheduledNotification> pendingNotifications =
                 scheduledNotificationRepository.findByTaskId(taskId);
